updating to 1.14

+++ src/main/java/com/storycraft/core/morph/entity/MorphEntity.java
@@ -2,8 +2,8 @@ package com.storycraft.core.morph.entity;
 
 import org.bukkit.entity.Player;
 
-import net.minecraft.server.v1_13_R2.DataWatcher;
-import net.minecraft.server.v1_13_R2.Entity;
+import net.minecraft.server.v1_14_R1.DataWatcher;
+import net.minecraft.server.v1_14_R1.Entity;
 
 public class MorphEntity implements IMorphEntity {
 
@@ -36,7 +36,7 @@ public class MorphEntity implements IMorphEntity {
     }
 
     @Override
-    public boolean onMoveSend(Player p, int deltaX, int deltaY, int deltaZ, boolean onGround) {
+    public boolean onMoveSend(Player p, short deltaX, short deltaY, short deltaZ, boolean onGround) {
         return false;
     }
 
@@ -46,7 +46,7 @@ public class MorphEntity implements IMorphEntity {
     }
 
     @Override
-    public boolean onLookAndMove(Player p, int deltaX, int deltaY, int deltaZ, byte yawAngle, byte pitchAngle,
+    public boolean onLookAndMove(Player p, short deltaX, short deltaY, short deltaZ, byte yawAngle, byte pitchAngle,
             boolean onGround) {
         return false;
     }
